backend/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
bandit==1.8.6
bcrypt==3.2.2
certifi==2025.8.3
cffi==2.0.0
click==8.2.1
colorama==0.4.6
coverage==7.10.7
cryptography==46.0.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
Mako==1.3.10
mando==0.7.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
nulltype==2.3.1
outcome==1.3.0.post0
packaging==25.0
passlib==1.7.4
plaid-python==36.1.0
pluggy==1.6.0
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.4.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
radon==6.0.1
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
ruff==0.13.0
sentry-sdk==2.37.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
sqlmodel==0.0.25
starlette==0.47.3
stevedore==5.5.0
trio==0.26.2
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
